Refactor: DRY up HTTP helpers, extract response.go, truncate errors
Security & code quality pass per review:
- Extract authQuery() to centralize credential extraction (single point)
- Extract sendAndDecode() to eliminate duplicated response handling
  between do() and doGet() (~50 lines of duplication removed)
- Move checkResponse to response.go (matches GitHub connector pattern)
- Add error message truncation (512 char cap) to prevent large Trello
  error bodies from bloating error payloads
- trello.go reduced from 309 to 258 lines
